+++ src/main/java/net/earlystage/compat/ExtraBlastingEmiRecipe.java
@@ -1,7 +1,5 @@
 package net.earlystage.compat;
 
-import java.util.List;
-
 import dev.emi.emi.api.recipe.EmiRecipe;
 import dev.emi.emi.api.recipe.EmiRecipeCategory;
 import dev.emi.emi.api.recipe.VanillaEmiRecipeCategories;
@@ -9,9 +7,13 @@ import dev.emi.emi.api.render.EmiTexture;
 import dev.emi.emi.api.stack.EmiIngredient;
 import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.api.widget.WidgetHolder;
+import net.earlystage.EarlyStageMain;
 import net.earlystage.misc.ExtraBlastingRecipe;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
+
+import java.util.List;
+import java.util.Locale;
 
 public class ExtraBlastingEmiRecipe implements EmiRecipe {
     private final List<EmiIngredient> input;
@@ -33,7 +35,7 @@ public class ExtraBlastingEmiRecipe implements EmiRecipe {
 
     @Override
     public Identifier getId() {
-        return null;
+        return EarlyStageMain.identifierOf("/" + this.output.get(0).getId().getPath() + "_extra");
     }
 
     @Override