+++ src/main/java/net/earlystage/init/EventInit.java
@@ -1,11 +1,11 @@
 package net.earlystage.init;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import ht.treechop.api.TreeChopEvents;
+import net.earlystage.EarlyStageMain;
+import net.earlystage.network.packet.SieveDropPacket;
 import net.fabricmc.fabric.api.event.registry.RegistryEntryAddedCallback;
 import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.item.AxeItem;
 import net.minecraft.item.Items;
@@ -16,6 +16,9 @@ import net.minecraft.loot.provider.number.BinomialLootNumberProvider;
 import net.minecraft.predicate.item.ItemPredicate;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class EventInit {
 
@@ -44,6 +47,10 @@ public class EventInit {
         if (FabricLoader.getInstance().isModLoaded("treechop")) {
             TreeChopEvents.BEFORE_CHOP.register((world, player, pos, state, chopData) -> player == null || player.getMainHandStack().getItem() instanceof AxeItem);
         }
+
+        ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
+            sender.sendPacket(new SieveDropPacket(EarlyStageMain.SIEVE_DROP_TEMPLATES));
+        });
     }
 
 }