+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Added:
 - Added easy respawn info
+- Added sieve emi compat
 ### Fixed:
-- 
+- Fixed blast furnace emi compat
 ### Changed:
 -